requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.2
idna==3.10
image==1.5.33
pillow==11.0.0
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
urllib3==2.2.3
